requirements.txt
scikit-learn>=1.3.0
joblib>=1.2.0S
catboost>=1.2
python-dotenv>=1.0.0
langchain>=0.0.249
openai>=0.27.8
streamlit-chat